refactor(lessonstable): 将公共组件中的LessonsTable组件改名lessonstable并放到pages/lessonstable下，并优化了一些代码规范问题
