+++ src/pages/lessonstable/index.module.scss
@@ -33,7 +33,7 @@
 }
 
 .conflict-time {
-  padding: 0 0 0 0;
+  padding: 0;
   width: 80%;
   margin: 0 auto;
   font-size: 1.3rem;
