+++ src/pages/lessonstable/utils/lesson-grid/dark.scss
