+++ src/pages/lessonstable/utils/lesson-grid/light.scss
